refactor RefreshDefinition 위치 변경 1. 공통으로 사용하는 정의에 대한 위치 변경
